requirements.txt
ipykernel
numpy
streamlit
transformers
python-dotenv
torch
tensorflow